+++ src/molecules/DayPickerUI.tsx
@@ -1,4 +1,39 @@
+import { useState } from 'react';
+import FocusTrap from 'focus-trap-react';
+import { DayPicker, SelectSingleEventHandler } from 'react-day-picker';
 import { PopperAtom } from '../atoms';
+
+interface IDayPickerUIProps {
+  showPopper: boolean;
+  popperRef: React.RefObject<HTMLDivElement>;
+  selected: Date;
+  handleClosePopper: () => void;
+  handleDaySelect: SelectSingleEventHandler;
+}
+
+const DayPickerUI = ({
+  showPopper,
+  popperRef,
+  selected,
+  handleClosePopper,
+  handleDaySelect,
+}: IDayPickerUIProps) => {
+  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(
+    null,
+  );
+
+  return (
+    <>
+      {showPopper && (
+        <FocusTrap
+          active
+          focusTrapOptions={{
+            initialFocus: false,
+            allowOutsideClick: true,
+            clickOutsideDeactivates: true,
+            onDeactivate: handleClosePopper,
+          }}
+        >
           <PopperAtom
             popperElement={popperElement}
             setPopperElement={setPopperElement}
@@ -9,4 +44,22 @@ import { PopperAtom } from '../atoms';
             role="dialog"
             ariaLabel="Daypicker calendar"
           >
+            <DayPicker
+              initialFocus={showPopper}
+              mode="single"
+              selected={selected}
+              onSelect={handleDaySelect}
+              defaultMonth={new Date()}
+              fromMonth={new Date()}
+              disabled={{ before: new Date() }}
+              required
+            />
           </PopperAtom>
+        </FocusTrap>
+      )}
+    </>
+  );
+};
+
+export default DayPickerUI;
+export { type IDayPickerUIProps };
