+++ src/molecules/index.ts
@@ -5,10 +5,9 @@ export {
 } from './LogInToUnlock';
 export { default as Clock } from './Clock';
 export { default as SideButtons } from './SideButtons';
-export { default as TodoCard } from './TodoCard';
 export { default as FlipCounter } from './FlipCounter';
 export { default as CurrentTodo } from './CurrentTodo';
 export { default as ExtremeModeIndicator } from './ExtremeModeIndicator';
 export { default as NowCard } from './NowCard';
-export { default as CalendarInput } from './CalendarInput';
 export { default as CategoryInput } from './CategoryInput';
+export { default as DayPickerUI, type IDayPickerUIProps } from './DayPickerUI';
