+++ src/organisms/index.ts
@@ -4,3 +4,4 @@ export { default as Records } from './Records';
 export type { IRecordsProps } from './Records';
 export { default as CurrentTodoCard } from './CurrentTodoCard';
 export { default as DateCard } from './DateCard';
+export { default as CalendarInput } from './CalendarInput';
